+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Date format: (YYYY-MM-DD)
 * Added: The give, transfer list and remove commands show the player's uuid as hover text now and allow it to be copied into the chat input via shift clicking.
 * Added: Debug option 'owner-name-updates' enables additional output whenever stored shop owner names get updated.
 * Added: Checking for invalid name-regex and then reverting back to default.
+* Fixed: Raider shopkeeper mobs were able to join and thereby interfere with nearby raids. This should no longer be the case.
 * Fixed: The save file was missing the data-version when initially created. This caused subsequent reloads to always trigger a 'migration' / forced without actually being required.
 * Fixed: The book shopkeeper was ignoring books with missing generation tag. These are now treated as 'original' books, just like Minecraft does.
 * Fixed: We would previously drop the shop-creation item returned on shop deletion at the shop's location, even if the shop got deleted via remote editing from far away (and is potentially not even loaded). If the player is further than 10 blocks away (or if the shop object is not loaded), it will drop the item at the player's location now.